Move github.com/ReconfigureIO/reco => pkg/reco
Part of an effort to make our main projects follow a standard go layout better
(needed for some vendoring tools to work optimally).
