+++ cmd/root.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 
 	"github.com/ReconfigureIO/cobra"
-	"github.com/ReconfigureIO/reco"
+	"github.com/ReconfigureIO/reco/pkg/reco"
 	"github.com/spf13/viper"
 )
